+++ properties/Adapter/AddDirectoryFromAnotherAdapter.php
@@ -9,7 +9,10 @@ use Innmind\Filesystem\{
     File,
     Name,
 };
-use Innmind\Immutable\Sequence;
+use Innmind\Immutable\{
+    Sequence,
+    SideEffect,
+};
 use Innmind\BlackBox\{
     Property,
     Set,
@@ -58,7 +61,9 @@ final class AddDirectoryFromAnotherAdapter implements Property
         );
 
         $assert->false($adapter->contains($directory->name()));
-        $assert->null($adapter->add($directory));
+        $assert
+            ->object($adapter->add($directory)->unwrap())
+            ->instance(SideEffect::class);
         $assert->true($adapter->contains($directory->name()));
         $assert->true(
             $adapter->get($directory->name())->match(