+++ properties/Adapter/RemoveFile.php
@@ -39,7 +39,9 @@ final class RemoveFile implements Property
 
     public function ensureHeldBy(Assert $assert, object $adapter): object
     {
-        $assert->null($adapter->add($this->file));
+        $assert
+            ->object($adapter->add($this->file)->unwrap())
+            ->instance(SideEffect::class);
         $assert->true($adapter->contains($this->file->name()));
         $assert
             ->object($adapter->remove($this->file->name())->unwrap())