+++ tests/Adapter/FilesystemTest.php
@@ -46,7 +46,12 @@ class FilesystemTest extends TestCase
 
         $this->assertInstanceOf(Adapter::class, $adapter);
         $this->assertFalse($adapter->contains(Name::of('foo')));
-        $this->assertNull($adapter->add(Directory::of(Name::of('foo'))));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $adapter
+                ->add(Directory::of(Name::of('foo')))
+                ->unwrap(),
+        );
         $this->assertTrue($adapter->contains(Name::of('foo')));
         $this->assertInstanceOf(
             SideEffect::class,
@@ -97,7 +102,7 @@ class FilesystemTest extends TestCase
                 Directory::of(Name::of('bar'))
                     ->add(File::of(Name::of('bar.md'), Content::ofString('# Bar'))),
             );
-        $adapter->add($directory);
+        $adapter->add($directory)->unwrap();
         $this->assertSame(
             '# Foo',
             $adapter
@@ -165,9 +170,9 @@ class FilesystemTest extends TestCase
 
         $d = Directory::of(Name::of('foo'));
         $d = $d->add(File::of(Name::of('bar'), Content::ofString('some content')));
-        $a->add($d);
+        $a->add($d)->unwrap();
         $d = $d->remove(Name::of('bar'));
-        $a->add($d);
+        $a->add($d)->unwrap();
         $this->assertSame(1, $d->removed()->count());
         $a = Filesystem::mount(Path::of('/tmp/'));
         $this->assertFalse(
@@ -187,10 +192,10 @@ class FilesystemTest extends TestCase
 
         $d = Directory::of(Name::of('foo'));
         $d = $d->add(File::of(Name::of('bar'), Content::ofString('some content')));
-        $a->add($d);
+        $a->add($d)->unwrap();
         $d = $d->remove(Name::of('bar'));
-        $a->add($d);
-        $a->add($d);
+        $a->add($d)->unwrap();
+        $a->add($d)->unwrap();
         $this->assertSame(1, $d->removed()->count());
         $a = Filesystem::mount(Path::of('/tmp/'));
         $this->assertFalse(
@@ -230,10 +235,12 @@ class FilesystemTest extends TestCase
     public function testRoot()
     {
         $adapter = Filesystem::mount(Path::of('/tmp/test/'));
-        $adapter->add(File::of(
-            Name::of('foo'),
-            Content::ofString('foo'),
-        ));
+        $adapter
+            ->add(File::of(
+                Name::of('foo'),
+                Content::ofString('foo'),
+            ))
+            ->unwrap();
         \file_put_contents('/tmp/test/bar', 'bar');
         \mkdir('/tmp/test/baz');
         \file_put_contents('/tmp/test/baz/foobar', 'baz');
@@ -294,8 +301,18 @@ class FilesystemTest extends TestCase
             Content::ofString('foo'),
         );
 
-        $this->assertNull($adapter->add($file));
-        $this->assertNull($adapter->add($file));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $adapter
+                ->add($file)
+                ->unwrap(),
+        );
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $adapter
+                ->add($file)
+                ->unwrap(),
+        );
     }
 
     public function testPathTooLongThrowAnException()
@@ -329,7 +346,7 @@ class FilesystemTest extends TestCase
                     ),
                 ),
             ),
-        ));
+        ))->unwrap();
     }
 
     public function testPersistedNameCanStartWithAnyAsciiCharacter()
@@ -348,15 +365,18 @@ class FilesystemTest extends TestCase
 
                 $filesystem = Filesystem::mount(Path::of($path));
 
-                $this->assertNull($filesystem->add(Directory::of(
-                    Name::of(\chr($ord).'a'),
-                    Sequence::of(
-                        File::of(
-                            Name::of('a'),
-                            Content::ofString($content),
+                $this->assertInstanceOf(
+                    SideEffect::class,
+                    $filesystem->add(Directory::of(
+                        Name::of(\chr($ord).'a'),
+                        Sequence::of(
+                            File::of(
+                                Name::of('a'),
+                                Content::ofString($content),
+                            ),
                         ),
-                    ),
-                )));
+                    ))->unwrap(),
+                );
 
                 (new FS)->remove($path);
             });
@@ -378,15 +398,18 @@ class FilesystemTest extends TestCase
 
                 $filesystem = Filesystem::mount(Path::of($path));
 
-                $this->assertNull($filesystem->add(Directory::of(
-                    Name::of('a'.\chr($ord).'a'),
-                    Sequence::of(
-                        File::of(
-                            Name::of('a'),
-                            Content::ofString($content),
+                $this->assertInstanceOf(
+                    SideEffect::class,
+                    $filesystem->add(Directory::of(
+                        Name::of('a'.\chr($ord).'a'),
+                        Sequence::of(
+                            File::of(
+                                Name::of('a'),
+                                Content::ofString($content),
+                            ),
                         ),
-                    ),
-                )));
+                    ))->unwrap(),
+                );
 
                 (new FS)->remove($path);
             });
@@ -410,15 +433,18 @@ class FilesystemTest extends TestCase
 
                 $filesystem = Filesystem::mount(Path::of($path));
 
-                $this->assertNull($filesystem->add(Directory::of(
-                    Name::of(\chr($ord)),
-                    Sequence::of(
-                        File::of(
-                            Name::of('a'),
-                            Content::ofString($content),
+                $this->assertInstanceOf(
+                    SideEffect::class,
+                    $filesystem->add(Directory::of(
+                        Name::of(\chr($ord)),
+                        Sequence::of(
+                            File::of(
+                                Name::of('a'),
+                                Content::ofString($content),
+                            ),
                         ),
-                    ),
-                )));
+                    ))->unwrap(),
+                );
 
                 (new FS)->remove($path);
             });