+++ tests/Adapter/LoggerTest.php
@@ -36,7 +36,12 @@ class LoggerTest extends TestCase
         );
         $file = File::of(Name::of('foo'), Content::none());
 
-        $this->assertNull($adapter->add($file));
+        $this->assertInstanceOf(
+            SideEffect::class,
+            $adapter
+                ->add($file)
+                ->unwrap(),
+        );
         $this->assertTrue($inner->contains($file->name()));
     }
 
@@ -48,7 +53,9 @@ class LoggerTest extends TestCase
         );
         $name = Name::of('foo');
         $file = File::of($name, Content::none());
-        $inner->add($file);
+        $inner
+            ->add($file)
+            ->unwrap();
 
         $this->assertSame(
             $file,
@@ -66,7 +73,9 @@ class LoggerTest extends TestCase
             new NullLogger,
         );
         $name = Name::of('foo');
-        $inner->add(File::of($name, Content::none()));
+        $inner
+            ->add(File::of($name, Content::none()))
+            ->unwrap();
 
         $this->assertTrue($adapter->contains($name));
     }
@@ -78,7 +87,9 @@ class LoggerTest extends TestCase
             new NullLogger,
         );
         $name = Name::of('foo');
-        $inner->add(File::of($name, Content::none()));
+        $inner
+            ->add(File::of($name, Content::none()))
+            ->unwrap();
 
         $this->assertInstanceOf(
             SideEffect::class,
@@ -99,7 +110,9 @@ class LoggerTest extends TestCase
             'watev',
             Content::none(),
         );
-        $inner->add($file);
+        $inner
+            ->add($file)
+            ->unwrap();
 
         $this->assertSame([$file], $adapter->root()->all()->toList());
     }